refactor(connectors): move share/actionable predicates into lib to end the layer inversion (R1-H3)
`lib/connectorPresentation` imported `isSharedConnector` from
`hooks/domain/useConnectorsController` — the only one of 36 files in `ui/src/lib`
importing from `hooks/`; the convention runs the other way (the hook imports
`connectorRowKey` from `lib`). It worked via tree-shaking, but any test importing
`connectorPresentation` dragged react-query + react into the graph.

Move the two pure predicates (`isSharedConnector`, `isActionableConnector`) into
`lib/connectorRows`; the hook re-exports them so the pages' import path is
unchanged. `lib/` now imports nothing from `hooks/`.

Class-check: `from '../hooks/` en ui/src/lib -> 1 sitio (connectorPresentation)
-> 0 tras el fix. Sin repro-test: refactor puro sin cambio de comportamiento,
cubierto por connectorPresentation.test / connectorRows.test / useConnectorsController.test (18 verdes).
